requirements.txt
django
django-extensions
djangorestframework